Fix for par 190 - there isn't a sample showModal method in github.
Added a sample showModal method to FormBase in the Erewhon version of AtcgGeneratorSchema.

Added an example of support for multiple user applications in the EDProfile::startup method of the Erewhon AtcgTestCodeSchema. Modified BasicDemo.cls to work with this, and to speed up replay.

Added another demo profile class ModalDemo.cls that demonstrates modal forms.